+++ packages/react-magma-dom/src/components/Breadcrumb/Item.tsx
@@ -31,7 +31,7 @@ const StyledSpan = styled.span<{ isInverse?: boolean }>`
       : props.theme.colors.neutral03};
 
   svg {
-    margin: 0 10px;
+    margin: 0 ${props => props.theme.spaceScale.spacing03};
   }
 `;