+++ packages/react-magma-dom/src/components/Select/Select.tsx
@@ -35,6 +35,7 @@ export function Select<T>(props: SelectInterface<T>) {
     onKeyDown,
     onKeyPress,
     onKeyUp,
+    placeholder,
     selectedItem: passedInSelectedItem,
   } = props;
 
@@ -144,6 +145,12 @@ export function Select<T>(props: SelectInterface<T>) {
     .replace(/\{labelText\}/g, labelText)
     .replace(/\{selectedItem\}/g, itemToString(selectedItem));
 
+  const selectText = itemToString(selectedItem)
+    ? itemToString(selectedItem)
+    : typeof placeholder === 'string'
+    ? placeholder
+    : i18n.select.placeholder;
+
   return (
     <SelectContainer
       getLabelProps={getLabelProps}
@@ -161,16 +168,14 @@ export function Select<T>(props: SelectInterface<T>) {
         isInverse={isInverse}
         style={inputStyle}
       >
-        <SelectText data-testid="selectedItemText">
-          {itemToString(selectedItem) || i18n.select.placeholder}
-        </SelectText>
+        <SelectText data-testid="selectedItemText">{selectText}</SelectText>
         {isClearable && selectedItem && (
           <ClearIndicator
             aria-label={clearIndicatorAriaLabel}
-            icon={<CrossIcon size={12} />}
+            icon={<CrossIcon size={10} />}
             onClick={defaultHandleClearIndicatorClick}
             size={ButtonSize.small}
-            style={{ margin: '0 4px', height: '28px', width: '28px' }}
+            style={{ marginTop: '0', marginBottom: '0' }}
             testId="clearIndicator"
             variant={ButtonVariant.link}
           />