+++ packages/react-magma-icons/CHANGELOG.md
@@ -3,13 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [2.0.0-beta.128](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.127...v2.0.0-beta.128) (2020-11-17)
+### [1.0.4](http://stash.cengage.com:7999///compare/react-magma-icons@1.0.3...react-magma-icons@1.0.4) (2020-11-24)
+
+## 2.0.0-beta.137 (2020-11-20)
 
 **Note:** Version bump only for package react-magma-icons
 
+# [2.0.0](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.136...v2.0.0) (2020-11-19)
 
+# <<<<<<< HEAD
 
+**Note:** Version bump only for package react-magma-icons
 
+# [2.0.0-beta.128](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.127...v2.0.0-beta.128) (2020-11-17)
+
+> > > > > > > master
+
+**Note:** Version bump only for package react-magma-icons
+
+# [2.0.0-beta.128](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.127...v2.0.0-beta.128) (2020-11-17)
+
+**Note:** Version bump only for package react-magma-icons
 
 # [2.0.0-beta.126](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.125...v2.0.0-beta.126) (2020-11-17)